Separating EthernetConnectionManager definition/implementation in a h/cpp file
